src/jquants_free_mcp_server/requirements.txt
httpx
FastMCP
streamlit
matplotlib
pandas
jquantsapi
python-dotenv